remove Validator.java

+++ src/main/java/jdbc/lesson4/homework/service/StorageService.java
@@ -15,32 +15,46 @@ public class StorageService {
     }
 
     public void delete(long id) throws InvalidInputException, SQLException {
-        Validator.checkId(id);
+        if (id < 0) {
+            throw new InvalidInputException("Id must be equals 0 or greater");
+        }
+
         STORAGE_DAO.delete(id);
     }
 
     public Storage update(Storage storage) throws InvalidInputException, SQLException {
         validateStorage(storage);
-
-        long freeSpace = STORAGE_DAO.getFreeSpace(STORAGE_DAO.findById(storage.getId()));
-        long newMaxSize = storage.getStorageMaxSize();
-        if (freeSpace > newMaxSize) {
-            throw new InvalidInputException("Unable to update storage(id=" + storage.getId() + ")." +
-                    " Old storage's free space(" + freeSpace + " can't be lower than new storage's max size("
-                    + newMaxSize + ")");
-        }
+        validateStorageNewFreeSpace(storage);
 
         return STORAGE_DAO.update(storage);
     }
 
     public Storage findById(long id) throws InvalidInputException, SQLException {
-        Validator.checkId(id);
+        if (id < 0) {
+            throw new InvalidInputException("Id must be equals 0 or greater");
+        }
         return STORAGE_DAO.findById(id);
     }
 
+    private void validateStorageNewFreeSpace(Storage storage) throws SQLException, InvalidInputException {
+        long freeSpace = STORAGE_DAO.getFreeSpace(STORAGE_DAO.findById(storage.getId()));
+        long newMaxSize = storage.getStorageMaxSize();
+
+        if (freeSpace > newMaxSize) {
+            throw new InvalidInputException("Unable to update storage(id=" + storage.getId() + ")." +
+                    " Old storage's free space(" + freeSpace + " can't be lower than new storage's max size("
+                    + newMaxSize + ")");
+        }
+    }
+
     void validateStorage(Storage storage) throws InvalidInputException {
-        Validator.checkForNull(storage);
-        Validator.checkId(storage.getId());
+        if (storage == null) {
+            throw new InvalidInputException("Input can't be null");
+        }
+
+        if (storage.getId() < 0) {
+            throw new InvalidInputException("Id must be equals 0 or greater");
+        }
 
         if (storage.getStorageMaxSize() <= 0) {
             throw new InvalidInputException("Size(" + storage.getStorageMaxSize() + ") must be greater than 0");
